requirements.txt
networkx==3.4.2
numpy==2.2.3
scipy==1.15.2
tkinterdnd2==0.4.2
pillow==11.1.0
